refactor: move various helpers from utils to utils-common
